+++ pages/dayTwo/index.js
@@ -1,5 +1,5 @@
 import { data } from "../../components/utils/data";
-import RoomNav from "../../components/RoomNav";
+import RoomNav from "../../components/TalksPerRoom";
 import Header from "../../components/Header";
 import StyledLink from "../../components/StyledLink";
